Update to the main storyboard and a new view controller.
