+++ tipify/tipViewController.swift
@@ -1,8 +1,8 @@
 //
-//  ViewController.swift
+//  tipViewController.swift
 //  tipify
 //
-//  Created by Nick McDonald on 12/15/16.
+//  Created by Nick McDonald on 12/16/16.
 //  Copyright © 2016 Nick McDonald. All rights reserved.
 //
 
@@ -10,14 +10,12 @@ import UIKit
 
 class tipViewController: UIViewController {
 
-    @IBOutlet weak var tipLabel: UILabel!
-    @IBOutlet weak var totalLabel: UILabel!
-    @IBOutlet weak var billTextField: UITextField!
-    @IBOutlet weak var tipSegmentedControl: UISegmentedControl!
+    @IBOutlet weak var totalBillTextField: UITextField!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
+
+        // Do any additional setup after loading the view.
     }
 
     override func didReceiveMemoryWarning() {
@@ -25,23 +23,20 @@ class tipViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
-    @IBAction func onTap(_ sender: Any) {
-        
-        view.endEditing(true)
-        
+    override func viewWillAppear(_ animated: Bool) {
+        // Make the textfield the first responder.
+        totalBillTextField.becomeFirstResponder()
     }
+    
 
-    @IBAction func calculateTip(_ sender: AnyObject) {
-        
-        let tipPercentages = [0.18, 0.2, 0.25]
-        
-        let billTotal = Double(billTextField.text!) ?? 0
-        let tip = billTotal*tipPercentages[tipSegmentedControl.selectedSegmentIndex]
-        let total = billTotal + tip
-        
-        tipLabel.text = String(format: "$%.2f", tip)
-        totalLabel.text = String(format: "$%.2f", total)
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
     }
+    */
 
 }
-
