+++ tipify/tpViewController.swift
@@ -1,0 +1,47 @@
+//
+//  ViewController.swift
+//  tipify
+//
+//  Created by Nick McDonald on 12/15/16.
+//  Copyright © 2016 Nick McDonald. All rights reserved.
+//
+
+import UIKit
+
+class tpViewController: UIViewController {
+
+    @IBOutlet weak var tipLabel: UILabel!
+    @IBOutlet weak var totalLabel: UILabel!
+    @IBOutlet weak var billTextField: UITextField!
+    @IBOutlet weak var tipSegmentedControl: UISegmentedControl!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+    @IBAction func onTap(_ sender: Any) {
+        
+        view.endEditing(true)
+        
+    }
+
+    @IBAction func calculateTip(_ sender: AnyObject) {
+        
+        let tipPercentages = [0.18, 0.2, 0.25]
+        
+        let billTotal = Double(billTextField.text!) ?? 0
+        let tip = billTotal*tipPercentages[tipSegmentedControl.selectedSegmentIndex]
+        let total = billTotal + tip
+        
+        tipLabel.text = String(format: "$%.2f", tip)
+        totalLabel.text = String(format: "$%.2f", total)
+    }
+
+}
+
